+++ src/helpers.cpp
@@ -94,7 +94,7 @@ void printTripleProgressBars(double progress1, double progress2, double progress
     std::cout.flush();
 }
 
-void printTripleProgressBars(double progress1, double progress2, double progress3, size_t a, size_t b, size_t c, size_t d, size_t e) {
+void printTripleProgressBars(double progress1, double progress2, double progress3, uint64_t a, uint64_t b, uint64_t c, uint64_t d, uint64_t e) {
     int barWidth = 70; // Width for each progress bar
     std::cout << "\033[3A"; // Move cursor up 3 lines
